+++ features/scripts/bundle_one_flavor.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+cd features/fixtures/app
+./gradlew :module:clean :module:bundleFoo --stacktrace