+++ src/main/groovy/com/bugsnag/android/gradle/BugsnagPlugin.groovy
@@ -169,7 +169,7 @@ class BugsnagPlugin implements Plugin<Project> {
         setupBugsnagTask(releasesTask, deps)
 
         if (shouldUploadDebugMappings(project, deps.output)) {
-            findAssembleTasks(deps.output, project).forEach {
+            findAssembleBundleTasks(deps.output, project).forEach {
                 releasesTask.mustRunAfter it
 
                 if (project.bugsnag.autoReportBuilds) {
@@ -190,7 +190,7 @@ class BugsnagPlugin implements Plugin<Project> {
         uploadTask.applicationId = deps.variant.applicationId
 
         if (shouldUploadDebugMappings(project, deps.output)) {
-            findAssembleTasks(deps.output, project).forEach {
+            findAssembleBundleTasks(deps.output, project).forEach {
                 uploadTask.mustRunAfter it
 
                 if (project.bugsnag.autoUpload) {
@@ -201,15 +201,32 @@ class BugsnagPlugin implements Plugin<Project> {
     }
 
     /**
-     * Fetches all the assemble tasks in the current project that match the variant
+     * Fetches all the assemble and bundle tasks in the current project that match the variant
      *
-     * Expected behaviour: assemble, assembleJavaExampleRelease, assembleJavaExample, assembleRelease
+     * Expected behaviour: [assemble, assembleJavaExampleRelease, assembleJavaExample, assembleRelease,
+     * bundle, bundleJavaExampleRelease, bundleJavaExample, bundleRelease]
      *
      * @param output the variantOutput
      * @param project the current project
-     * @return the assemble tasks
+     * @return the assemble + bundle tasks
      */
-    private static Set<Task> findAssembleTasks(BaseVariantOutput output, Project project) {
+    private static Set<Task> findAssembleBundleTasks(BaseVariantOutput output, Project project) {
+        Set<String> taskNames = new HashSet<>()
+        taskNames.addAll(findTaskNamesForPrefix(output, "assemble"))
+        taskNames.addAll(findTaskNamesForPrefix(output, "bundle"))
+
+        project.tasks.findAll {
+            taskNames.contains(it.name)
+        }
+    }
+
+    /**
+     * Finds all the task names which can be used to assemble a variant, and replaces 'assemble' with the given
+     * prefix.
+     *
+     * E.g. [bundle, bundleRelease, bundleFooRelease]
+     */
+    private static Set<String> findTaskNamesForPrefix(BaseVariantOutput output, String prefix) {
         String variantName = output.name.split("-")[0].capitalize()
         String assembleTaskName = output.assemble.name
         String buildTypeTaskName = assembleTaskName.replaceAll(variantName, "")
@@ -217,14 +234,11 @@ class BugsnagPlugin implements Plugin<Project> {
         String variantTaskName = assembleTaskName.replaceAll(buildType, "")
 
         Set<String> taskNames = new HashSet<>()
-        taskNames.add(assembleTaskName)
-        taskNames.add("assemble")
-        taskNames.add(buildTypeTaskName)
-        taskNames.add(variantTaskName)
-
-        project.tasks.findAll {
-            taskNames.contains(it.name)
-        }
+        taskNames.add(prefix)
+        taskNames.add(assembleTaskName.replaceAll("assemble", prefix))
+        taskNames.add(buildTypeTaskName.replaceAll("assemble", prefix))
+        taskNames.add(variantTaskName.replaceAll("assemble", prefix))
+        return taskNames
     }
 
     private static void setupManifestUuidTask(Project project, BugsnagTaskDeps deps) {