+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 3.X.X (TBD)
 
+* Support automatic upload when building App Bundles
+[#137](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/137)
+
+### Bug fixes
+
 * Create tasks for debug variant which are not run by default
 [#139](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/139)